requirements.txt
dm-sonnet==1.36
google-pasta==0.2.0
h5py==3.1.0
imageio==2.9.0
jupyter==1.0.0
kiwisolver==1.3.1
Markdown==3.3.4
MarkupSafe==1.1.1
matplotlib==3.3.4
numpy==1.19.5
pandas==1.1.5
PyWavelets==1.1.1
scikit-image==0.17.2
scipy==1.5.4
semantic-version==2.8.5
tensorboard==1.15.0
tensorflow==1.15.0
tensorflow-estimator==1.15.1
tensorflow-gpu==1.15.0
tensorflow-probability==0.8.0
termcolor==1.1.0
tifffile==2020.9.3
torchfile==0.1.0
